reset only if rw exists

+++ packtivity/statecontexts/posixfs_context.py
@@ -48,7 +48,7 @@ class LocalFSState(object):
         resets state by deleting readwrite directory contents (deletes tree and re-creates)
         '''
         for rw in self.readwrite + [self.metadir]:
-            if os.path.exists(rw):
+            if rw and os.path.exists(rw):
                 shutil.rmtree(rw)
         self.ensure()
